Update index.html
- Added more information to the warning cards used in the No Compatible Version Found section.

Now shows the followers and the tag icon that when hovered over shows the all the categories the mod has listed. Crucially now shows a Reason that when clicked toggles a drop down, same way the changelog does for the success cards, that gives you the reason why this mod is not compatible.