fix: fixed added searched mods not clearing
Fixed mods that were added via search not properly clearing when clearing the search box of text.

Changed the Clear Input button to also clear the search box as well.

Set the search modal to open and focus on the search bar when the user clicks into the search box on the main page.