feat: Enhanced searching to include "Featured" mods
The search modal now includes "Featured" mods that will show when the user first clicks the search box or clears any text within the search box. It will randomly choose an option in the sort by and the project type drop down menus so it feels a little more dynamic each time the user uses the search. You can still manually use these drop downs and search for mods using them. The drop down menus will be hidden when the user manually searches for a mod by name.

Re-styled the search cards to feel and look more like the other cards in the ap.

Filtering the category tags to only include the actual mod categories and not things like "Fabric, Iris"...etc.